login,sigin,session,cookie

+++ src/app/controllers/CourseController.js
@@ -1,17 +1,50 @@
 const Course = require('../models/Course');
 const { mutipleMongooseToObject } = require('../../util/mongoose');
-const {mongooseToObject} = require('../../util/mongoose');
+const { mongooseToObject } = require('../../util/mongoose');
+const PAGE_SIZE = 9;
 class CourseController {
 
     //[GET]/course
-    showAll(req, res,next){
-        Course.find({})
-            .then(courses => {
-                res.render('courses/courses', { 
-                    courses: mutipleMongooseToObject(courses)
-                });
-            })
-            .catch(next);
+    showAll(req, res, next) {
+        var page = req.query.page;
+        page = parseInt(page);
+        if (page) {
+            //get of page
+            if(page<1){
+                page=1;
+            }
+            var skiper = (page-1)*PAGE_SIZE;
+            Course.find({})
+                .skip(skiper)
+                .limit(PAGE_SIZE)
+                .then(courses => {
+                    Course.countDocuments({})
+                        .then((total)=>{
+                            // console.log(total);
+                            var totalPage = Math.ceil(total/PAGE_SIZE);
+                            const data = courses;
+                            //console.log(totalPage);
+                            //console.log(data);
+                            res.render('courses/courses', {
+                                courses: mutipleMongooseToObject(courses),
+                                totalPage: totalPage,
+                                total: total,
+                                data:data
+                            });
+                        });
+                    
+                })
+                .catch(next);
+        } else {
+            //get all courses
+            Course.find({})
+                .then(courses => {
+                    res.render('courses/courses', {
+                        courses: mutipleMongooseToObject(courses)
+                    });
+                })
+                .catch(next);
+        }
     }
     //[GET]/course/:slug
     showItems(req, res, next) {
@@ -23,11 +56,17 @@ class CourseController {
     }
     //[GET]/course/create
     create(req, res, next) {
+        if (!req.session.isAuthenticated){
+            return res.redirect('/users/sign-in');
+        }
         res.render('courses/create');
     }
     //[GET]/course/store
     store(req, res, next) {
-        req.body.image = `https://img.youtube.com/vi/${req.body.linkImg}/maxresdefault.jpg`
+        if (!req.session.isAuthenticated){
+            return res.redirect('/users/sign-in');
+        }
+        req.body.userId = req.session.authUser.email;
         const course = new Course(req.body);
         course.save()
             .then(() => res.redirect('/me/stored/courses'))
@@ -37,6 +76,9 @@ class CourseController {
     }
     //[GET]/course/:id/edit
     edit(req, res, next) {
+        if (!req.session.isAuthenticated){
+            return res.redirect('/users/sign-in');
+        }
         Course.findById(req.params.id)
             .then(course => res.render('courses/edit', {
                 course: mongooseToObject(course)
@@ -45,30 +87,45 @@ class CourseController {
     }
     //[PUT]/course/:id
     update(req, res, next) {
+        if (!req.session.isAuthenticated){
+            return res.redirect('/users/sign-in');
+        }
         Course.updateOne({ _id: req.params.id }, req.body)
             .then(() => res.redirect('/me/stored/courses'))
             .catch(next);
     }
     //[DELETE]/course/:id
     destroy(req, res, next) {
+        if (!req.session.isAuthenticated){
+            return res.redirect('/users/sign-in');
+        }
         Course.delete({ _id: req.params.id })
-            .then(() => res.redirect('back'))
-            .catch(next);
-    }
-    //[PATCH]/course/:id/restore
-    restore(req, res, next) {
-        Course.restore({ _id: req.params.id })
             .then(() => res.redirect('back'))
             .catch(next);
     }
     //[DELETE]/course/:id/force
     forceDestroy(req, res, next) {
+        if (!req.session.isAuthenticated){
+            return res.redirect('/users/sign-in');
+        }
         Course.deleteOne({ _id: req.params.id })
+            .then(() => res.redirect('back'))
+            .catch(next);
+    }
+    //[PATCH]/course/:id/restore
+    restore(req, res, next) {
+        if (!req.session.isAuthenticated){
+            return res.redirect('/users/sign-in');
+        }
+        Course.restore({ _id: req.params.id })
             .then(() => res.redirect('back'))
             .catch(next);
     }
     //[POST]/course/handle-form
     handleForm(req, res, next) {
+        if (!req.session.isAuthenticated){
+            return res.redirect('/users/sign-in');
+        }
         switch (req.body.action) {
             case 'delete':
                 Course.delete({ _id: { $in: req.body.courseIds } })
@@ -81,6 +138,9 @@ class CourseController {
     }
     //[POST]/course/trash-all
     trashAll(req, res, next) {
+        if (!req.session.isAuthenticated){
+            return res.redirect('/users/sign-in');
+        }
         switch (req.body.action) {
             case 'deleteForce':
                 Course.deleteOne({ _id: { $in: req.body.courseIds } })
@@ -95,10 +155,6 @@ class CourseController {
             default:
                 res.json('Không có hành vi');
         }
-    }
-    //[GET]/course/:slug/learn
-    learn(req, res, next) {
-        res.render('courses/learn');
     }
 }
 //Public ra ngoài
